requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/dev.in
#
asgiref==3.7.2
    # via django
astroid==2.15.5
    # via pylint
black==23.3.0
    # via -r requirements/dev.in
click==8.1.3
    # via black
dill==0.3.6
    # via pylint
django==4.2.2
    # via
    #   -r requirements/base.in
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.14.0
    # via
    #   -r requirements/base.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements/base.in
environs==9.5.0
    # via -r requirements/base.in
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via
    #   -r requirements/dev.in
    #   pylint
lazy-object-proxy==1.9.0
    # via astroid
marshmallow==3.19.0
    # via environs
mccabe==0.7.0
    # via pylint
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via
    #   black
    #   marshmallow
    #   pytest
pathspec==0.11.1
    # via black
platformdirs==3.8.0
    # via
    #   black
    #   pylint
pluggy==1.2.0
    # via pytest
psycopg2-binary==2.9.6
    # via -r requirements/base.in
pyjwt==2.7.0
    # via djangorestframework-simplejwt
pylint==2.17.4
    # via -r requirements/dev.in
pytest==7.4.0
    # via
    #   -r requirements/dev.in
    #   pytest-django
    #   pytest-env
    #   pytest-mock
pytest-django==4.5.2
    # via -r requirements/dev.in
pytest-env==0.8.2
    # via -r requirements/dev.in
pytest-mock==3.11.1
    # via -r requirements/dev.in
python-dotenv==1.0.0
    # via environs
pytz==2023.3
    # via djangorestframework
ruff==0.0.275
    # via -r requirements/dev.in
sqlparse==0.4.4
    # via django
tomlkit==0.11.8
    # via pylint
wrapt==1.15.0
    # via astroid
